requirements.txt
Django==5.2.5
djangorestframework==3.15.2
paytechuz==0.3.40
environs==11.0.0
requests==2.32.4
django-unfold==0.74.1
